+++ src/app/core/dispatch.service.js
@@ -52,10 +52,15 @@ console.info(key, keys);
           chartService.getCharts(key);
           break;
         case 'hd-styles':
-          var srtr = key.split(':');
-          var style = {}; style[srtr[0]] = srtr[1];
-console.warn(style);
+          var srtr = key.split(':'), style = {}
+          style[srtr[0]] = srtr[1];
           handsontableService.addSelectedAreaStyle(style);
+          break;
+        case 'hd-floatSize': handsontableService.adjustFloatSize(key); break;
+        case 'hd-calc':
+          var srtr = key.split(':'), calc = {}
+          calc[srtr[0]] = (srtr[1] === 'true');
+          handsontableService.addSelectedAreaCalc(calc);
           break;
         default: break;
       }